Move document fetching to lib
