+++ pages/tags/[tag].tsx
@@ -4,9 +4,9 @@ import React from 'react'
 
 import ContentCardList from '../../components/ContentCardList'
 import PageLayout from '../../components/PageLayout'
+import loadAllRecords from '../../lib/loadAllRecords'
 import collectTags from '../../util/collectTags'
 import getQueryParameter from '../../util/getQueryParameters'
-import loadAllRecords from '../../util/loadAllRecords'
 import { MarkdownFileWithUrl } from '../../util/MarkdownFile'
 import omitUndefinedFields from '../../util/omitUndefinedFields'
